YAML Configuration Support
Added the \Fgms\SiteErrorNotifications\YamlFactory class which transforms a YAML configuration document into a \Fgms\SiteErrorNotifications\ErrorHandlerInterface object.

Integrated the aforementioned class into the WordPress plugin.  This means that a configuration file must be provided.  By default it is expected to be "config.yml" in the root directory of the plugin.  See comment in site-error-notifications.php for details on how to change this path.