At Operator Support
The PHP @ operator suppresses PHP errors for the statement to which it is applied however as per http://php.net/manual/en/function.set-error-handler.php the error handler is still run.

Added the \Fgms\SiteErrorNotifications\AtOperatorErrorHandler class (and supporting base classes) which decorates an inner instance of \Fgms\SiteErrorNotifications\ErrorHandlerInterface and only invokes it if the @ operator has not been used to suppress the current error.

The \Fgms\SiteErrorNotifications\YamlFactory class now wraps the \Fgms\SiteErrorNotifications\CompositeErrorHandler it generates in an \Fgms\SiteErrorNotifications\AtOperatorErrorHandler.

Closes #6.