\Fgms\SiteErrorNotifications\Utility::getBacktrace Fix/Improvement
When PHP invokes a custom error handler that stack frame in a backtrace doesn't have an associated file or line due to the fact PHP is invoking it internally (it's not invoked from anywhere in the code it's invoked from within the engine itself therefore there is no file or line).  This broke code which relied on all stack frames having an associated file and line.

Modified \Fgms\SiteErrorNotifications\Utility::getBacktrace to use this criterion (rather than hacky is_subclass_of tests) to determine where to cut off the backtrace.